Version 0.9.4. This version has a 'working' version of Rally/Track mode. This has not been tested yet, will be tested soon and bugs will be fixed soon.

+++ CruisePW.h
@@ -3,7 +3,7 @@
 class Cruise : public Mode
 {
   private:
-    const int minSpeed = 30;
+    const byte minSpeed = 30;
     const float baseVoltage50 = 1.13;
 
     float currentVoltage = 0;
@@ -11,7 +11,7 @@ class Cruise : public Mode
     short averageSpeedSize = 10;
     float averageSpeed[10];
     float currentSpeed = 0;
-    int desiredSpeed = -1;
+    short desiredSpeed = -1;
     bool enabled = false;
     bool increaseDesiredSpeed = false;
     byte speedChangeFeedbackIndex = 0;
